requirements.txt
# Instagram API
instagrapi==2.1.5

# Image processing
Pillow==10.1.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# AI APIs
openai==1.3.0
google-generativeai==0.3.0
anthropic==0.7.0

# Environment management
python-dotenv==1.0.0

# Additional utilities
lxml==4.9.3